Added Package Priority logic and rider assign to packages
Added logic to prioritize the Packages based on conditions into priority queue.
Added assign rider to packages logic to utilize the riders in the most efficient way

+++ src/main/java/com/chronos/couriers/model/riderinfo/Rider.java
@@ -9,6 +9,7 @@ public class Rider {
     private final int reliabilityRating;
     private final boolean canHandleFragile;
 
+
     public Rider(String riderId, String riderName, int reliabilityRating, boolean canHandleFragile) {
         this.riderId = riderId;
         this.riderName = riderName;
@@ -47,11 +48,12 @@ public class Rider {
         return maxLoad;
     }
 
-    public double getReliabilityRating() {
+    public int getReliabilityRating() {
         return reliabilityRating;
     }
 
     public boolean isCanHandleFragile() {
         return canHandleFragile;
     }
+
 }
